Fixed a few warnings.
Rev: src/array.c:1.83
Rev: src/backend.c:1.54
Rev: src/docode.c:1.78
Rev: src/encode.c:1.66
Rev: src/error.c:1.60
Rev: src/error.h:1.48
Rev: src/errors.h:1.12
Rev: src/fsort.c:1.14
Rev: src/hashtable.c:1.7
Rev: src/main.c:1.96
Rev: src/pike_memory.h:1.21
Rev: src/pike_types.c:1.133
Rev: src/pike_types.h:1.43

+++ src/error.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: error.h,v 1.47 2000/08/15 11:05:19 grubba Exp $
+ * $Id: error.h,v 1.48 2000/08/15 16:11:50 grubba Exp $
  */
 #ifndef ERROR_H
 #define ERROR_H
@@ -59,8 +59,8 @@ typedef struct JMP_BUF
   struct JMP_BUF *previous;
   jmp_buf recovery;
   struct pike_frame *frame_pointer;
-  INT32 stack_pointer;
-  INT32 mark_sp;
+  ptrdiff_t stack_pointer;
+  ptrdiff_t mark_sp;
   INT32 severity;
   ONERROR *onerror;
 #ifdef PIKE_DEBUG
